refactor: renamed <tool-tip> to <app-tooltip>
