+++ src/heads-up/components/index.js
@@ -4,4 +4,4 @@ export { default as PanelSection } from './panel-section.vue'
 export { default as PropertiesList } from './properties-list.vue'
 export { default as RefreshButton } from './refresh-button.vue'
 export { default as ResourceList } from './resource-list.vue'
-export { default as ToolTip } from './tool-tip.vue'
+export { default as AppTooltip } from './app-tooltip.vue'
